+++ outputs.tf
@@ -10,16 +10,3 @@ output "subnetwork" {
   value       = try(google_compute_subnetwork.subnetwork, null)
   description = "The created subnet resource"
 }
-
-# ------------------------------------------------------------------------------
-# OUTPUT ALL INPUT VARIABLES
-# ------------------------------------------------------------------------------
-
-# ------------------------------------------------------------------------------
-# OUTPUT MODULE CONFIGURATION
-# ------------------------------------------------------------------------------
-
-output "module_enabled" {
-  description = "Whether the module is enabled."
-  value       = var.module_enabled
-}