+++ test/unit-disabled/main.tf
@@ -1,19 +1,3 @@
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-# EMPTY FEATURES (DISABLED) UNIT TEST
-# This module tests an empty set of features.
-# The purpose is to verify no resources are created when the module is disabled.
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-terraform {
-  required_providers {
-    aws = {
-      source  = "hashicorp/google"
-      version = "4.12.0"
-    }
-  }
-}
-
-# DO NOT RENAME MODULE NAME
 module "test" {
   source = "../.."
 
@@ -33,9 +17,3 @@ module "test" {
     }
   ]
 }
-
-# outputs generate non-idempotent terraform plans so we disable them for now unless we need them.
-# output "all" {
-#   description = "All outputs of the module."
-#   value       = module.test
-# }