+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- BREAKING CHANGE: remove output `module_enabled`
+- BREAKING CHANGE: remove support for `google` provider version `3`
+
 ## [0.0.2]
 
 ### Fixed